netty_1

+++ src/main/java/wjt/netty/b/EchoClientHandler.java
@@ -19,7 +19,7 @@ public class EchoClientHandler implements ChannelInboundHandler {
     public EchoClientHandler() {
         firstMsg= Unpooled.buffer(1024);
         for (int i=0;i<firstMsg.capacity();i++){
-            firstMsg.writeInt(i);
+            firstMsg.writeByte((byte)i);
         }
 
     }
